Separate VerbnetFrame into official and instance
VerbnetFrame was cut in three:
  * ComputeSlotTypeMixin used by both, infers things like "subject" or
    "object" for arguments
  * VerbnetFrameOccurrence: FrameNet frames converted to VerbNet
  * VerbnetOfficialFrame: VerbNet frames from the resources

The goal is to reduce confusion: a VerbnetFrame could be two very
different things conceptually. It also was an issue because some members
were only used in one of the two cases.

However, the code still contains "frame" variables: eliminate them to
produce frame_occurrence or verbnet_frame?