Simplify FrameNet -> VerbNet conversion
The previous algorithm operated by doing "inplace" string replacements,
but this is a poor fit in Python: strings are immutable. It's also a bit
low-level: we need to keep count of offsets, and so on.

The new algorithm builds a list of dictionaries containing all necessary
information to go back to the 'NP V NP' form, is simpler, and can
accommodate further modifications, such as storing the preposition
inside the PP instead of doing prep NP.

The behavior is not exactly the same: some weird corner cases got
removed, but the results should be 99% the same.